fix(phase5): Add type casting for ROW_NUMBER() in get_object_timeline()
Fixed DatatypeMismatch error where ROW_NUMBER() returns BIGINT but
get_object_timeline() function signature expects INTEGER for timeline_version.

Added ::INTEGER cast to match Phase 4 pattern and function return type.

Test Results:
- Phase 5: Improved from 19/37 to 26/37 passing (70% ↑)
- TestGetObjectTimeline: All 8 tests now passing
- Remaining 11 failures in get_audit_trail() and query_at_timestamp() functions
  which have separate SQL issues (ambiguous columns, missing functions)

Overall Combined: 53/70 tests passing (76%)

+++ sql/033_pggit_history_audit.sql
@@ -420,7 +420,7 @@ BEGIN
     RETURN QUERY
     WITH object_history_seq AS (
         SELECT
-            ROW_NUMBER() OVER (ORDER BY oh.created_at ASC) AS timeline_version,
+            ROW_NUMBER() OVER (ORDER BY oh.created_at ASC)::INTEGER AS timeline_version,
             oh.change_type,
             oh.before_definition,
             oh.after_definition,
